feat(phaseG): nodes RAM residency — bounds trim + role-split caches + PIR/pin cleanup
Phase G RAM-residency bundle (nodes side, #53 lever 2 + audit items B/C/D/K/L):

- RECV_QUEUE_SIZE 8 -> 4 (Mesh.h): ~1KB static RAM cut, the dominant lever in
  this bundle (observed fan-in never exceeded 3 concurrent frames).
- LATTICE_ROUTE_TABLE_MAX 32 -> 16 (master-only heap allocation, Phase B).
- LATTICE_REPLAY_MAX_ORIGINS 16 -> 12.
- Audit B: LATTICE_E2E_KEYCACHE_MAX role split. E2EKeyStore's backing array is
  now heap-allocated (mirrors RouteTable's Phase B role-conditional pattern)
  and shrinks to 2 slots for leaves via Mesh::reevaluateRouteTable(), vs. 10
  for masters (~568B/leaf saved) — a real allocation shrink, not just a
  logical cap.
- Audit C: Enrollment::PENDING_RELAY_QUEUE_SIZE 8 -> 4, tracking RECV_QUEUE_SIZE.
- Audit D: ReplayCache::Entry field reorder for padding (24B -> 20B/entry).
- Audit K: PirAdapter's _cooldownSeconds -> static constexpr; _timerActive/
  _motionSent collapsed into one PirState enum.
- Audit L: Adapter::_pin int -> uint8_t, propagated through AdapterFactory and
  Adapter/PirAdapter/SerialAdapter constructors.
- New CompactMessage.{h,cpp} (§7): a correct, tested wire<->compact converter
  (212B vs wire's 250B) — NOT wired into Mesh::recvQueue. Traced every
  recvQueue-reachable consumer of the fields the design doc's sketch proposed
  dropping and confirmed (including empirically, via two e2e test failures
  that were reverted) that route_path/auth_tag/auth_path/enrollment_public_key
  and even the JOIN_ACK secondary-master fields are load-bearing against the
  current (pre-Task-4) wire format's relay/verify paths. Full rationale in
  task-3-report.md. Forward-looking: usable once Task 4 moves the JOIN_ACK
  secondary fields into data[64].

223/223 unit tests + 41/41 e2e tests pass.

Part of lattice-nodes Phase G (issue #53).

+++ firmware/main/src/adapter/Adapter.cpp
@@ -14,7 +14,7 @@ namespace adapter {
 
 using namespace lattice::utils;
 
-Adapter::Adapter(int pin)
+Adapter::Adapter(uint8_t pin)
     : _pin(pin), _adapterType(adapter_types::UNKNOWN_ADAPTER), mesh_transmit_fn(nullptr) {
   LATTICE_LOGLN("Adapter", "Base adapter initialized with UNKNOWN_ADAPTER", LogLevel::LOG_DEBUG);
 }
